+++ sentry-rails/spec/dummy/test_rails_app/public/sentry-logo.png
